setup App class
