+++ include/App/App.h
@@ -1,0 +1,17 @@
+#pragma once
+
+class App
+{
+public:
+    App();
+    void setup();
+    void loop();
+};
+
+
+
+
+
+
+
+
